+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## [5.1.1](https://github.com/untemps/svelte-palette/compare/v5.1.0...v5.1.1) (2026-03-26)
+
+### Bug Fixes
+
+-   Restore drag placeholder visibility in drop deletion mode ([#145](https://github.com/untemps/svelte-palette/issues/145)) ([93fc18c](https://github.com/untemps/svelte-palette/commit/93fc18c86d98fedd294a975f3076e1759c57c119))
+
+# [5.1.0](https://github.com/untemps/svelte-palette/compare/v5.0.1...v5.1.0) (2026-03-26)
+
+### Features
+
+-   Add maxColumns prop to limit columns in auto-expanding mode ([#142](https://github.com/untemps/svelte-palette/issues/142)) ([efa4c1f](https://github.com/untemps/svelte-palette/commit/efa4c1fb933df4424afd10a59e097039bfc3a678))
+
+## [5.0.1](https://github.com/untemps/svelte-palette/compare/v5.0.0...v5.0.1) (2026-03-26)
+
+### Bug Fixes
+
+-   Normalize rgb/rgba output from EyeDropper API ([#141](https://github.com/untemps/svelte-palette/issues/141)) ([509bc38](https://github.com/untemps/svelte-palette/commit/509bc38655e141a9369593479b4112b2f1c05239))
+
 # [5.0.0](https://github.com/untemps/svelte-palette/compare/v4.1.0...v5.0.0) (2026-03-16)
 
 -   feat!: Migrate to Svelte 5 ([e9551d6](https://github.com/untemps/svelte-palette/commit/e9551d60ad7d717be8dac94d4719adbe0914672c))